cannon: Remove support for singlethreaded and 32-bit multithreaded cannon versions.
