+++ cannon/mipsevm/memory/memory64_benchmark_test.go
@@ -1,6 +1,3 @@
-//go:build cannon64
-// +build cannon64
-
 package memory
 
 import (
