+++ cannon/mipsevm/tests/fuzz_evm_common64_test.go
@@ -1,6 +1,3 @@
-//go:build cannon64
-// +build cannon64
-
 package tests
 
 import (
