+++ cannon/mipsevm/tests/helpers.go
@@ -12,21 +12,10 @@ import (
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/arch"
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/multithreaded"
 	mttestutil "github.com/ethereum-optimism/optimism/cannon/mipsevm/multithreaded/testutil"
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/singlethreaded"
-	sttestutil "github.com/ethereum-optimism/optimism/cannon/mipsevm/singlethreaded/testutil"
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/testutil"
 )
 
 type VMFactory func(po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger, opts ...testutil.StateOption) mipsevm.FPVM
-
-func singleThreadedVmFactory(po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger, opts ...testutil.StateOption) mipsevm.FPVM {
-	state := singlethreaded.CreateEmptyState()
-	mutator := sttestutil.NewStateMutatorSingleThreaded(state)
-	for _, opt := range opts {
-		opt(mutator)
-	}
-	return singlethreaded.NewInstrumentedState(state, po, stdOut, stdErr, nil)
-}
 
 func multiThreadedVmFactory(po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger, features mipsevm.FeatureToggles, opts ...testutil.StateOption) mipsevm.FPVM {
 	state := multithreaded.CreateEmptyState()
@@ -39,13 +28,6 @@ func multiThreadedVmFactory(po mipsevm.PreimageOracle, stdOut, stdErr io.Writer,
 
 type ElfVMFactory func(t require.TestingT, elfFile string, po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger) mipsevm.FPVM
 
-func singleThreadElfVmFactory(t require.TestingT, elfFile string, po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger) mipsevm.FPVM {
-	state, meta := testutil.LoadELFProgram(t, elfFile, singlethreaded.CreateInitialState, true)
-	fpvm := singlethreaded.NewInstrumentedState(state, po, stdOut, stdErr, meta)
-	require.NoError(t, fpvm.InitDebug())
-	return fpvm
-}
-
 func multiThreadElfVmFactory(t require.TestingT, elfFile string, po mipsevm.PreimageOracle, stdOut, stdErr io.Writer, log log.Logger, features mipsevm.FeatureToggles) mipsevm.FPVM {
 	state, meta := testutil.LoadELFProgram(t, elfFile, multithreaded.CreateInitialState, false)
 	fpvm := multithreaded.NewInstrumentedState(state, po, stdOut, stdErr, log, meta, features)
@@ -54,20 +36,6 @@ func multiThreadElfVmFactory(t require.TestingT, elfFile string, po mipsevm.Prei
 }
 
 type ProofGenerator func(t require.TestingT, state mipsevm.FPVMState, memoryProofAddresses ...arch.Word) []byte
-
-func singleThreadedProofGenerator(t require.TestingT, state mipsevm.FPVMState, memoryProofAddresses ...arch.Word) []byte {
-	var proofData []byte
-
-	insnProof := state.GetMemory().MerkleProof(state.GetPC())
-	proofData = append(proofData, insnProof[:]...)
-
-	for _, addr := range memoryProofAddresses {
-		memProof := state.GetMemory().MerkleProof(addr)
-		proofData = append(proofData, memProof[:]...)
-	}
-
-	return proofData
-}
 
 func multiThreadedProofGenerator(t require.TestingT, state mipsevm.FPVMState, memoryProofAddresses ...arch.Word) []byte {
 	mtState, ok := state.(*multithreaded.State)
@@ -97,18 +65,6 @@ type VersionedVMTestCase struct {
 	Version        versions.StateVersion
 }
 
-func GetSingleThreadedTestCase(t require.TestingT) VersionedVMTestCase {
-	return VersionedVMTestCase{
-		Name:           "single-threaded",
-		Contracts:      testutil.TestContractsSetup(t, testutil.MipsSingleThreaded, 0),
-		StateHashFn:    singlethreaded.GetStateHashFn(),
-		VMFactory:      singleThreadedVmFactory,
-		ElfVMFactory:   singleThreadElfVmFactory,
-		ProofGenerator: singleThreadedProofGenerator,
-		Version:        versions.VersionSingleThreaded2,
-	}
-}
-
 func GetMultiThreadedTestCase(t require.TestingT, version versions.StateVersion) VersionedVMTestCase {
 	features := versions.FeaturesForVersion(version)
 	return VersionedVMTestCase{
@@ -126,23 +82,12 @@ func GetMultiThreadedTestCase(t require.TestingT, version versions.StateVersion)
 	}
 }
 
-func GetMultiThreadedTestCases(t require.TestingT) []VersionedVMTestCase {
+func GetMipsVersionTestCases(t require.TestingT) []VersionedVMTestCase {
 	var cases []VersionedVMTestCase
 	for _, version := range versions.StateVersionTypes {
-		if arch.IsMips32 && versions.IsSupportedMultiThreaded(version) {
-			cases = append(cases, GetMultiThreadedTestCase(t, version))
-		}
 		if !arch.IsMips32 && versions.IsSupportedMultiThreaded64(version) {
 			cases = append(cases, GetMultiThreadedTestCase(t, version))
 		}
-	}
-	return cases
-}
-
-func GetMipsVersionTestCases(t require.TestingT) []VersionedVMTestCase {
-	cases := GetMultiThreadedTestCases(t)
-	if arch.IsMips32 {
-		cases = append(cases, GetSingleThreadedTestCase(t))
 	}
 	return cases
 }
