+++ op-challenger/game/fault/trace/cannon/provider_test.go
@@ -11,12 +11,12 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/multithreaded"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/stretchr/testify/require"
 
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/memory"
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/singlethreaded"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/trace/utils"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/trace/vm"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
@@ -51,7 +51,7 @@ func TestGet(t *testing.T) {
 
 	t.Run("ProofAfterEndOfTrace", func(t *testing.T) {
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &singlethreaded.State{
+		generator.finalState = &multithreaded.State{
 			Memory: memory.NewMemory(),
 			Step:   10,
 			Exited: true,
@@ -107,7 +107,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("GenerateProof", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &singlethreaded.State{
+		generator.finalState = &multithreaded.State{
 			Memory: memory.NewMemory(),
 			Step:   10,
 			Exited: true,
@@ -133,7 +133,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("ProofAfterEndOfTrace", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &singlethreaded.State{
+		generator.finalState = &multithreaded.State{
 			Memory: memory.NewMemory(),
 			Step:   10,
 			Exited: true,
@@ -159,7 +159,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("ReadLastStepFromDisk", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, initGenerator := setupWithTestData(t, dataDir, prestate)
-		initGenerator.finalState = &singlethreaded.State{
+		initGenerator.finalState = &multithreaded.State{
 			Memory: memory.NewMemory(),
 			Step:   10,
 			Exited: true,
@@ -177,7 +177,7 @@ func TestGetStepData(t *testing.T) {
 		require.Contains(t, initGenerator.generated, 7000, "should have tried to generate the proof")
 
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &singlethreaded.State{
+		generator.finalState = &multithreaded.State{
 			Memory: memory.NewMemory(),
 			Step:   10,
 			Exited: true,
@@ -250,7 +250,7 @@ func setupWithTestData(t *testing.T, dataDir string, prestate string) (*CannonTr
 
 type stubGenerator struct {
 	generated  []int // Using int makes assertions easier
-	finalState *singlethreaded.State
+	finalState *multithreaded.State
 	proof      *utils.ProofData
 
 	finalStatePath string
